Move helm repo function (valve on -> valve toolbox install)

+++ src/common.sh
@@ -132,14 +132,16 @@ _select_one() {
 }
 
 _config_save() {
-    FIND_CNT=$(cat ${CONFIG} | grep REGISTRY= | wc -l | xargs)
-    if [ ${FIND_CNT} -eq 0 ]; then
-        echo "# valve config" > ${CONFIG}
-        echo "REGISTRY=${REGISTRY:-docker-registry.127.0.0.1.nip.io:30500}" >> ${CONFIG}
-        echo "CHARTMUSEUM=${CHARTMUSEUM:-chartmuseum-devops.coruscant.opsnow.com}" >> ${CONFIG}
-        echo "USERNAME=${USERNAME}" >> ${CONFIG}
-    else
-        _result "CONFIG Set"
+    if [ -f ${CONFIG} ]; then
+        FIND_CNT=$(cat ${CONFIG} | grep REGISTRY= | wc -l | xargs)
+        if [ ${FIND_CNT} -eq 0 ]; then
+            echo "# valve config" > ${CONFIG}
+            echo "REGISTRY=${REGISTRY:-docker-registry.127.0.0.1.nip.io:30500}" >> ${CONFIG}
+            echo "CHARTMUSEUM=${CHARTMUSEUM:-chartmuseum-devops.coruscant.opsnow.com}" >> ${CONFIG}
+            echo "USERNAME=${USERNAME}" >> ${CONFIG}
+        else
+            _result "CONFIG Set"
+        fi
     fi
 }
